fix: MachException Improvements (#2662)
Follow up on https://github.com/getsentry/sentry-cocoa/pull/2642. Fixes applied from KSCrash:
* kstenerud/KSCrash#349
* kstenerud/KSCrash#350

+++ Sources/SentryCrash/Recording/SentryCrashReport.c
@@ -1289,7 +1289,7 @@ writeError(const SentryCrashReportWriter *const writer, const char *const key,
                 writer->addStringElement(writer, SentryCrashField_CodeName, machCodeName);
             }
             writer->addUIntegerElement(
-                writer, SentryCrashField_Subcode, (unsigned)crash->mach.subcode);
+                writer, SentryCrashField_Subcode, (size_t)crash->mach.subcode);
         }
         writer->endContainer(writer);
 #endif
